add other ios devices

+++ scripts/ios_integration_test.sh
@@ -1,19 +1,29 @@
 #!/bin/bash
+set -euxo pipefail
 
-flutter pub get
+if [ "$RECORD" == "true" ]
+then
+  # due to a bug in xcode we need to sleep a while, else the recording is empty
+  xcrun simctl io booted recordVideo --codec=h264 --mask=black recording.mov & sleep 5
 
-# due to a bug in xcode need to sleep a while, else the recording is empty
-
-xcrun simctl io booted recordVideo recording.mov & sleep 5
-
-export RECORDING_PID=${!}
-echo "Recording process up with pid: ${RECORDING_PID}"
+  export RECORDING_PID=${!}
+  echo "Recording process up with pid: ${RECORDING_PID}"
+fi
 
 flutter drive --target=test_driver/app.dart
+flutter drive --target=test_driver/scanPage.dart
 
-sleep 5
-kill -SIGINT $RECORDING_PID
-sleep 10
+mkdir -p "$TEMP_DIR"
 
-mkdir "$TEMP_DIR"
-cp recording.mov "$TEMP_DIR"
+if [ "$RECORD" == "true" ]
+then
+  sleep 5
+  kill -SIGINT $RECORDING_PID
+  sleep 10
+  cp recording.mov "$TEMP_DIR"
+fi
+
+
+# copy screenshots to TEMP_DIR
+cp -r "$TEMP_DIR/test"/* "$TEMP_DIR"
+rm -r "$TEMP_DIR/test"
